fix(web): update provider status when API key is cleared
- Updated `handleKeyChange` in `page.tsx` and `settings/page.tsx` to trim values and remove keys from state when empty.
- Updated `resolveKeySource` in `lib/keys.ts` to trim keys before checking truthiness.
- Updated `handleHistoryLoad` in `page.tsx` to close mobile menu on selection for better UX and test reliability.
- Added `web/tests/e2e/key-management.spec.ts` with mobile/tablet support to verify the fix.
- Reverted unintentional changes to `web/next-env.d.ts`.
